Implement index mutation
