+++ pkg/v1/mutate/mutate.go
@@ -17,8 +17,6 @@ package mutate
 import (
 	"archive/tar"
 	"bytes"
-	"encoding/json"
-	"errors"
 	"fmt"
 	"io"
 	"io/ioutil"
@@ -28,8 +26,6 @@ import (
 
 	v1 "github.com/google/go-containerregistry/pkg/v1"
 	"github.com/google/go-containerregistry/pkg/v1/empty"
-	"github.com/google/go-containerregistry/pkg/v1/partial"
-	"github.com/google/go-containerregistry/pkg/v1/stream"
 	"github.com/google/go-containerregistry/pkg/v1/tarball"
 	"github.com/google/go-containerregistry/pkg/v1/types"
 	"github.com/google/go-containerregistry/pkg/v1/v1util"
@@ -69,6 +65,25 @@ func Append(base v1.Image, adds ...Addendum) (v1.Image, error) {
 		base: base,
 		adds: adds,
 	}, nil
+}
+
+type Appendable interface {
+	MediaType() (types.MediaType, error)
+	Digest() (v1.Hash, error)
+	Size() (int64, error)
+}
+
+type IndexAddendum struct {
+	Add Appendable
+	v1.Descriptor
+}
+
+// AppendManifest appends a manifest to tbe ImageIndex.
+func AppendManifests(base v1.ImageIndex, adds ...IndexAddendum) v1.ImageIndex {
+	return &index{
+		base: base,
+		adds: adds,
+	}
 }
 
 // Config mutates the provided v1.Image to have the provided v1.Config
@@ -112,227 +127,6 @@ func CreatedAt(base v1.Image, created v1.Time) (v1.Image, error) {
 	cfg.Created = created
 
 	return ConfigFile(base, cfg)
-}
-
-type image struct {
-	base v1.Image
-	adds []Addendum
-
-	computed   bool
-	configFile *v1.ConfigFile
-	manifest   *v1.Manifest
-	diffIDMap  map[v1.Hash]v1.Layer
-	digestMap  map[v1.Hash]v1.Layer
-}
-
-var _ v1.Image = (*image)(nil)
-
-func (i *image) MediaType() (types.MediaType, error) { return i.base.MediaType() }
-
-func (i *image) compute() error {
-	// Don't re-compute if already computed.
-	if i.computed {
-		return nil
-	}
-	var configFile *v1.ConfigFile
-	if i.configFile != nil {
-		configFile = i.configFile
-	} else {
-		cf, err := i.base.ConfigFile()
-		if err != nil {
-			return err
-		}
-		configFile = cf.DeepCopy()
-	}
-	diffIDs := configFile.RootFS.DiffIDs
-	history := configFile.History
-
-	diffIDMap := make(map[v1.Hash]v1.Layer)
-	digestMap := make(map[v1.Hash]v1.Layer)
-
-	for _, add := range i.adds {
-		diffID, err := add.Layer.DiffID()
-		if err != nil {
-			return err
-		}
-		diffIDs = append(diffIDs, diffID)
-		history = append(history, add.History)
-		diffIDMap[diffID] = add.Layer
-	}
-
-	m, err := i.base.Manifest()
-	if err != nil {
-		return err
-	}
-	manifest := m.DeepCopy()
-	manifestLayers := manifest.Layers
-	for _, add := range i.adds {
-		d := v1.Descriptor{}
-		var err error
-
-		if d.Size, err = add.Layer.Size(); err != nil {
-			return err
-		}
-
-		if d.Digest, err = add.Layer.Digest(); err != nil {
-			return err
-		}
-
-		if d.MediaType, err = add.Layer.MediaType(); err != nil {
-			return err
-		}
-
-		d.Annotations = add.Annotations
-		d.URLs = add.URLs
-
-		manifestLayers = append(manifestLayers, d)
-		digestMap[d.Digest] = add.Layer
-	}
-
-	configFile.RootFS.DiffIDs = diffIDs
-	configFile.History = history
-
-	manifest.Layers = manifestLayers
-
-	rcfg, err := json.Marshal(configFile)
-	if err != nil {
-		return err
-	}
-	d, sz, err := v1.SHA256(bytes.NewBuffer(rcfg))
-	if err != nil {
-		return err
-	}
-	manifest.Config.Digest = d
-	manifest.Config.Size = sz
-
-	i.configFile = configFile
-	i.manifest = manifest
-	i.diffIDMap = diffIDMap
-	i.digestMap = digestMap
-	i.computed = true
-	return nil
-}
-
-// Layers returns the ordered collection of filesystem layers that comprise this image.
-// The order of the list is oldest/base layer first, and most-recent/top layer last.
-func (i *image) Layers() ([]v1.Layer, error) {
-	if err := i.compute(); err == stream.ErrNotComputed {
-		// Image contains a streamable layer which has not yet been
-		// consumed. Just return the layers we have in case the caller
-		// is going to consume the layers.
-		layers, err := i.base.Layers()
-		if err != nil {
-			return nil, err
-		}
-		for _, add := range i.adds {
-			layers = append(layers, add.Layer)
-		}
-		return layers, nil
-	} else if err != nil {
-		return nil, err
-	}
-
-	diffIDs, err := partial.DiffIDs(i)
-	if err != nil {
-		return nil, err
-	}
-	ls := make([]v1.Layer, 0, len(diffIDs))
-	for _, h := range diffIDs {
-		l, err := i.LayerByDiffID(h)
-		if err != nil {
-			return nil, err
-		}
-		ls = append(ls, l)
-	}
-	return ls, nil
-}
-
-// ConfigName returns the hash of the image's config file.
-func (i *image) ConfigName() (v1.Hash, error) {
-	if err := i.compute(); err != nil {
-		return v1.Hash{}, err
-	}
-	return partial.ConfigName(i)
-}
-
-// ConfigFile returns this image's config file.
-func (i *image) ConfigFile() (*v1.ConfigFile, error) {
-	if err := i.compute(); err != nil {
-		return nil, err
-	}
-	return i.configFile, nil
-}
-
-// RawConfigFile returns the serialized bytes of ConfigFile()
-func (i *image) RawConfigFile() ([]byte, error) {
-	if err := i.compute(); err != nil {
-		return nil, err
-	}
-	return json.Marshal(i.configFile)
-}
-
-// Digest returns the sha256 of this image's manifest.
-func (i *image) Digest() (v1.Hash, error) {
-	if err := i.compute(); err != nil {
-		return v1.Hash{}, err
-	}
-	return partial.Digest(i)
-}
-
-// Size implements v1.Image.
-func (i *image) Size() (int64, error) {
-	if err := i.compute(); err != nil {
-		return -1, err
-	}
-	return partial.Size(i)
-}
-
-// Manifest returns this image's Manifest object.
-func (i *image) Manifest() (*v1.Manifest, error) {
-	if err := i.compute(); err != nil {
-		return nil, err
-	}
-	return i.manifest, nil
-}
-
-// RawManifest returns the serialized bytes of Manifest()
-func (i *image) RawManifest() ([]byte, error) {
-	if err := i.compute(); err != nil {
-		return nil, err
-	}
-	return json.Marshal(i.manifest)
-}
-
-// LayerByDigest returns a Layer for interacting with a particular layer of
-// the image, looking it up by "digest" (the compressed hash).
-func (i *image) LayerByDigest(h v1.Hash) (v1.Layer, error) {
-	if cn, err := i.ConfigName(); err != nil {
-		return nil, err
-	} else if h == cn {
-		return partial.ConfigLayer(i)
-	}
-	if layer, ok := i.digestMap[h]; ok {
-		return layer, nil
-	}
-	return i.base.LayerByDigest(h)
-}
-
-// LayerByDiffID is an analog to LayerByDigest, looking up by "diff id"
-// (the uncompressed hash).
-func (i *image) LayerByDiffID(h v1.Hash) (v1.Layer, error) {
-	if layer, ok := i.diffIDMap[h]; ok {
-		return layer, nil
-	}
-	return i.base.LayerByDiffID(h)
-}
-
-func validate(adds []Addendum) error {
-	for _, add := range adds {
-		if add.Layer == nil {
-			return errors.New("unable to add a nil layer to the image")
-		}
-	}
-	return nil
 }
 
 // Extract takes an image and returns an io.ReadCloser containing the image's
